move map into its own subcommand file
